Tweaked intro
Jason Perlow suggested this rewrite:

> This guide is intended to help open source projects, and maintainers create and maintain a coordinated vulnerability response process.

The comma insertion is not correct, but does suggest confusion.
I think this document is squarely directed to
"open source project maintainers" so I rewrote it that way.
It also impacts reporters/finders, but I while they are impacted,
the whole document is written specifically for maintainers.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>