Jason Perlow: Use 1 space (not 2) after sentences
The document inconsistently uses 1 space after sentences & in other
places it uses 2 spaces. This changes a paragraph from
2-space to 1-space separators.

I (David) prefer extra space between sentences. The extra space
helps readers realize that there's a new paragraph, and we want to
help readers. Tools like TeX and LaTex make this easy (and the default).

HOWEVER, this is a Markdown text that will often be viewed in HTML.
It's more complicated in Markdown and HTML to have
multiple space separators between sentences.
The easy way is to insert "\&nbsp; " after each non-final sentence; that
copies & pastes well, but is ugly in the edited Markdown. As noted in
<https://hea-www.harvard.edu/~fine/Tech/html-sentences.html>,
"The correct way to solve this problem turns out to be to use the
CSS word-spacing parameter, which does just what it sounds like it
does (more or less). So we set the word-spacing for paragraphs to
the space we want between sentences, and then set the word-spacing
in sentences back to normal. Perhaps not the most intuitive approach,
but it seems to work correctly in modern browsers."
But this requires CSS somewhere and spans for each sentence.
That's more complex.

In the interest of simplicity we'll just switch to 1-space
separators between sentences.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>