Release MWF - not (say) Fridays!
The log4j fix was released on a Friday, which created
a lot of problems for overworked admins over the weekend.
Maybe that was unavoidable, but it makes it clear that the
day of week of a release matters.

It turns out that worldwide, the weekend (non-working week)
is typically Friday, Saturday, and/or Sunday:
https://en.wikipedia.org/wiki/Workweek_and_weekend

Let's recommend security releases occur MWF, and explain why.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>